requirements.txt
streamlit
langchain-community
langchain-groq
langchain-chroma
langchain-huggingface
langchain-classic
pypdf
sentence-transformers
python-dotenv